fixed navigation and states causing bugs with data storage in room db (redo because files were not added)

+++ app/src/main/java/com/keith/expensesplitter/ui/fragments/MakeExpenseFragment.kt
@@ -77,8 +77,10 @@ class MakeExpenseFragment : Fragment() {
     private fun savingExpenses(expenses: List<ExpensesAdapter.ExpenseView>) {
         lifecycleScope.launch {
             expenses.forEach { expenseView ->
-                viewModel.makeExpense(expenseView.name,
-                    expenseView.amount)
+                viewModel.makeExpense(
+                    expenseView.name,
+                    expenseView.amount.toLong()
+                )
             }
         }
     }
